Return correct type from DeepSkyObject::getRenderMask()
More correctly is to return Renderer::RenderFlags, but this require big
render.h refactoring.

+++ src/celengine/globular.cpp
@@ -494,7 +494,7 @@ void Globular::renderGlobularPointSprites(
     glEnd();
 }
 
-unsigned int Globular::getRenderMask() const
+uint64_t Globular::getRenderMask() const
 {
     return Renderer::ShowGlobulars;
 }
